gitserver: Fix lazy repository initialization
Lazy init would not happen when refs would be fetched, meaning that this
endpoint would return an error and leave no chance for git to continue.
We now initialize an empty repository on first fetch/push and remove the
code that initially populates it, this is now left to clients.