adminserver: Intercept the auth cookie manually in the client
We can't rely on the underlying HTTP client of the RPC client to give
us back the auth cookie in full form, since http.CookieJar will only
return the cookie name and value, and we need all the fields to generate
a Netscape-format cookie file that curl can use (the cookie file is
passed to git for clone/push operations).

The old code used a simpler file format with a single Set-Cookie line
per cookie without domain or expiry date, but this simpler format
doesn't seem to work on all curl versions. It does work with the curl
shipped by Fedora, but doesn't with the curl shipped by Alpine.